Add permissions and fix some bugs
Add permission restrictions for certain commands and certain parts
of a command.

For example, if I want to set the channel in a command where there
are no subcommands, the user's permissions are checked in permissions_handler.

However, if there's a command that's completely restricted to a certain
permission, we use serenity's inbuilt framework to check the permission.

In addition, Serenity currently has a bug where parsing an argument
into a ChannelId also works with a UserId and therefore breaks the
storage mechanism. Instead, use parse_channel along with casting
the argument to a String.

Signed-off-by: kingbri <bdashore3@gmail.com>