Fix code formatting and add an argument limit
Use ctx as a reference when saying something in a channel and
add a limit to textmod commands because if someone puts an empty command,
an error will be fired and the bot will panic.

However, we cannot differentiate errors based on command name, so an
issue has been filed in the serenity repository to ask for that feature.

Signed-off-by: kingbri <bdashore3@gmail.com>