Initial commit using Twilight
Build a skeleton for the command_handler, credentials_helper, and
various structures for command contexts and data.

Serenity's hand-holding approach is not for me. Therefore, I'm using
Twilight which is the discord api skeleton and I am free to customize
whatever I want in CourtJester all the way down to the handler.

The bot will be rebuilt to its current state based on twilight rather
than serenity.

Signed-off-by: kingbri <bdashore3@gmail.com>