Upgrade the framework
Use Erk's context model instead of what I had before. This allows
us to grab a context as a reference instead of directly grabbing the type.

Also, remove the context creation function from command_utils and instead
use an impl to provide a create function.

Finally, use the Message model type instead of the MessageCreate type because
MessageCreate is a tuple that contains a Message.

Signed-off-by: kingbri <bdashore3@gmail.com>