Port over nice and bruh commands
For these commands to work, I have to do two things:
1. Infer if the user's going to put a channel as the first argument
2. Make sure a moderator or greater permission can set the message
channel

Using Results allows me to work with Errors rather than an arbitrary
number representing a false case. This allows for the bot to not
fail on a case where the ID is actually -1.

The Permissions handler system pulls user permissions from given
roles and adds them into an enum which is then compared in a match
statement to see if the user has the permission or not.

Signed-off-by: kingbri <bdashore3@gmail.com>