Add custom prefix support and perform updates
In my other bots, I've stored the custom prefixes in an I/O layer
that is used instead of direct database calls. Since this bot isn't
targeted towards a lot of servers, I am scrapping the need for
a Dashmap until it's needed.

To reduce the strain on database calls, only deal with the prefix
to check if the message should be passed on to the command handler AND
remove the prefix from the command itself.

This caused a change in the split_string function. It now includes
the prefix within the &str Vector. However, this doesn't pose a problem
since the command is only checked/used once. The other times this vector
is used is when we need a subcommand or to join a message.

In addition, change the Message type in command args to &Message due to various
move errors. Doing this allows for code consistency.

Signed-off-by: kingbri <bdashore3@gmail.com>