Add custom commands framework
Users can now add simple custom commands. User inputs a command trigger
and message to be sent. The message currently has one variable ({user})
which is replaced by the user's mention string on call. More will be added
as demand increases.

Also, update the twilight package to the latest commit which involved
some changes with Gateway Intents from the Discord API.

In addition, modify some of the string_renderer functions to be dynamic
based on joining strings and to work properly with custom prefixes

Signed-off-by: kingbri <bdashore3@gmail.com>