Add starbot functionality
starbot is when a message is posted and users react with the
star emoji. If there's a certain amount of stars, the message
is posted in a channel.

However, this function is similar to quoting and is redundant
when used together. Therefore, if someone enables starbot, quoting
is turned off for the guild. The server can have one or the other,
but not both.

When adding a reaction, if the star count is the same as the threshold,
a new message is created with the message info, the ID and the star count.

This updates based on the amount of stars given. If the star count
is less than the threshold (either by removing a reaction or an admin
clearing reactions), the message is deleted.

Rather than making another function for when stars are removed (due
to the different events), introduce a removal boolean that allows
for less LOC

I will introduce a second commit that will handle cases where the starbot
message is deleted and make the configuration prettier by using a wizard
for initialization.

Signed-off-by: kingbri <bdashore3@gmail.com>