Add image posting and nsfw check to starboard
I'll change all related file-names in another commit.

Starboard now has support for posting images with the extensions:
- gif
- png
- jpeg & jpg
- webp

Which are most commonly used in discord images.

In addition, if a channel is NSFW and the starboard channel isn't
NSFW, the post isn't starred and a warning is given to the user.

Also, fix a bug where messages containg only images errored out the bot

Signed-off-by: kingbri <bdashore3@gmail.com>