Add starboard reaction functionality
When someone reacts to a message with the star emoji, a message is posted
in a designated channel (Channel for the quote command if the user used
that before) with the id, star count, and message itself.

If there are more reactions, the message is edited accordingly. If there
are no more stars on the message or the star count is less than the threshold,
the message is deleted.

Messages are only posted if the star count passes the configured guild threshold

Starboard commands and setup wizard will be present in the next commit

Signed-off-by: kingbri <bdashore3@gmail.com>