Add starboard command
Configures the starboard, can only be used by administrators of the bot.

The starboard can either be setup using the wizard which leads the user
through a series of messages/reactions to get the configuration finished.
The threshold and channel can be manually set using the respective commands.

In addition, perform miscellaneous updates on the bot for clarity

Signed-off-by: kingbri <bdashore3@gmail.com>