Add music features
This commit adds music functionality using lavalink to the bot.

Commands:
- Play
- Pause
- Resume
- Stop
- Queue
- Now_playing

The queue fetches all tracks and places them in a list along
with the ones that are now playing. All other commands function the
same as other music bots since we're using lavalink to delegate
what music to play.

Upcoming feature is to find a YouTube URL using search keywords and
an idle disconnect so the bot doesn't use a lot of RAM being idle
on a Voice Channel.

Signed-off-by: kingbri <bdashore3@gmail.com>