Add idle timer for voice connection
Having the bot in the voice call for a long period of time can consume
a lot of system resources. This commit is able to mitigate that using
a simple timer.

By splitting the disconnect command and its internals, we now have access
to disconnect from the voice channel anywhere in the code.

On commands that are meant to make the bot not speak, a new abortable
future is created with a delay of 10 minutes (Subject to change).

The reason for an abortable future is if the bot needs to speak again,
but the disconnect delay is running. The kill switch is stored in a dashmap
categorized by guild id. Simply call the kill switch and the future is
never executed.

I will add support for idling when the queue is empty in the next commit.

Signed-off-by: kingbri <bdashore3@gmail.com>