Update music framework
Update from serenity-lavalink to the new lavalink-rs crate. This
implementation is simpler and easier to work with.

Idle timers weren't being properly disposed of and this causes the queue
garbage collector to prematurely return since the hashmap still contained
the aborthandler.

A planned feature is to add the time remaining on the current song when
using the queue command and to allow for mm:ss time when using seek.

Signed-off-by: kingbri <bdashore3@gmail.com>