Actually fix the starboard error
The starboard reaction error referenced in commit 91cdc4a was not the
actual fix.

I forgot to put a WHERE clause in the starboard itself which caused
the bot to fetch the first database entry that matched the join.

In addition, the wizard is now updated to where both row checking
and quote channel checking exist to get a more accurate result when
setting a starboard channel.

Signed-off-by: kingbri <bdashore3@gmail.com>