Separate workflows
There are three different tasks that need to be done by Actions
on a push, pull-request, and a release.

On a push, the bot is built and uploaded to the server.
On a pull-request, the bot is built as a check before merging
On a release, the bot is built and published as a linux binary

For now, there are no windows/darwin binaries since this bot was
made on linux and will require further testing for these OSes

Signed-off-by: kingbri <bdashore3@gmail.com>