Update dependencies
serenity-await, lavalink-rs, and sqlx are all published on crates.io
and are now reflected as such in the cargo.toml.

In addition, fix a bug in music where stopping the player doesn't
empty the queue by skipping the song and then stopping the player.

When a new song is played, it plays without hesitiation and doesn't
make the bot crash.

Signed-off-by: kingbri <bdashore3@gmail.com>