Fix event mismatch for guild_delete
For some reason, the guild_delete event no longer signifies when
a bot is kicked/banned from a guild.

Instead it uses the guild_member_remove event which is triggered every
time a member leaves.

For this reason, the bot id has to be stored and when a member leaves
the guild, the bot checks for self and deletes the record from the database.

Signed-off-by: kingbri <bdashore3@gmail.com>