Update permissions and errors
Add a custom error enum for CourtJester so it's easier to write when
an error happens.

Also use RoyalGuard's updated permissions handler and modify it
so it only checks for manage_messages.

In addition, dispatch errors weren't firing due to an error in serenity's
current release candidate, so switch to the current branch where
this fix is merged. I will switch back to the rcs once a new one
with the fix is released.

Since lavalink-rs requires serenity's rc, force the dependency to use
the given one in cargo

Signed-off-by: kingbri <bdashore3@gmail.com>