Add uwu command
Original from nitsuga5124: https://gitlab.com/nitsuga5124/robo-arc/-/blob/master/src/commands/fun.rs#L623

This improves on the original version by reducing the amount of mutable
variables and cascading .replace() instead of setting the mutable string
using multiple .replace() statements.

I could have used a map and matched the values of each word, but,
the replace statements are tailored in a specific order, so this method
is not viable.

Signed-off-by: kingbri <bdashore3@gmail.com>