Improve music and prefixes
Apparently, I misunderstod how dynamic prefixes work. They do not
default to the standard prefix if the prefix doesn't exist in the
cache. Instead, they work in conjunction with the hardcoded prefix.

Remove the hardcoded prefix and only use the dynamic prefix function.
If the prefix isn't found in the prefix map, use the default prefix.

Also, with music, two changes have been added:

1. Splice out a track from the queue: If a user accidentally adds
a track, the earlier option was to clear the entire queue. However,
support to remove the certain track from the queue is now available.
Providing the track number within the queue using the clear command
will remove that specific track.

2. Seeking with the hh:mm:ss format. If mm:ss or ss is provided,
the bot will take those values as such. If you want to seek to a specific
time, be sure to enter it in mm:ss format, otherwise the playhead will
jump to the seconds value of a singlular number!

Signed-off-by: kingbri <bdashore3@gmail.com>