Update error message and mentions
On a bot mention, the user accesses a list of "emergency commands"
which are run to get the bot back on track in terms of the prefix.

The bot should not be ran using a mention prefix and there are
safeguards against that in the new before hook.

Also, if there's a random bot error, it's posted on the discord
channel where the error occurs so users can forward errors to me
as needed.

Signed-off-by: kingbri <bdashore3@gmail.com>