Add anime/manga search and fix formatting
Anime and manga search has been added within the bot. This uses the
jikan api to grab the information of an anime/manga based on the title.

The engine gives results back and the user can choose which item he/she
wants to see. The original result list is deleted once a selection is made.

If a selection isn't made and 30 seconds pass, timeout the collector.

Also fix some odds and ends in terms of code format

Signed-off-by: kingbri <bdashore3@gmail.com>