starboard: Various fixes and optimizations
Fix: Reactions weren't properly checked. If someone reacted with
an emoji on the message before the star, that number was checked
instead of the star amount.

This was because the find() closure wasn't checking on the given
argument and instead checked on the reaction itself.

Also, make the starboard wizard/deactivate commands use as_data()
to fetch the emoji instead of a helper function.

Signed-off-by: kingbri <bdashore3@gmail.com>