Starboard: Add a temporary workaround for spam
Reaction spamming is a big problem when it comes to this starboard.
Since the tasks on reactions are executed in parallel, there can
be an issue in the form of multiple posts.

The temporary fix is to check if the message exists in the DB and if not,
add it to the starboard channel.

References #21
Resolves #20

Signed-off-by: kingbri <bdashore3@gmail.com>