Voice: Overhaul music security
The bot has a lot of exploits pertaining to voice channel security.
This can be annoying especially when a troll is playing around with
the music bot.

- Make sure the user and CourtJester are in the same voice channel
before executing any music command
- If a user tries to summon CourtJester in a different voice channel,
force the user to disconnect the bot before moving it to a different
channel.
- Clear the music queue and event loops on disconnect since these
are not cleared by default.

A possible future update will be starting a disconnect timer
upon joining the voice chat.

Signed-off-by: kingbri <bdashore3@gmail.com>