Reaction Images: Don't post the same gif twice
This QOL enhancement ensures that the same gif won't be used when
you use any reaction command.

Includes:
- hug
- pat
- slap
- cry
- cringe

This is achieved by loading the last used image URL into the cache
and then checking if that image URL exists on the next fetch of gifs.
The mechanism uses a dashmap to preserve concurrency and thread safety.

Signed-off-by: kingbri <bdashore3@gmail.com>