Voice utils: Delete already awaited futures
On timed bot disconnects, a future is delayed and later awaited to
disconnect CourtJester from the voice chat.

When using an abortable future (the core of this timer logic), the
abort handler is stored in a dashmap with an identifying key.

When the abort handler is called to abort the future, the entry in
the dashmap isn't deleted.

This commit deletes the dashmap entry after the future has been awaited.

To see why I did this, reference bdashore3/RoyalGuard@befe3a111e62830e9aebb42ec4b60a90333eeceb

Signed-off-by: kingbri <bdashore3@gmail.com>